i think this is actually working now
caveats:

no nesting in the middle of a class name with &, e.g.:

  .colorSection {
    &-title {
    }
  }

behavior of globals is a little off maybe? i had to refactor the use of
‘sheet’, ‘sheet-title’, etc. to be non-global

for some reason the webpack dev server throws errors about the use of
anonymous styleName in a context with multiple stylesheet imports… even
when there aren’t multiple stylesheet imports. so i always do named
imports with “eslint-disable-line no-unused-vars”

instead of “composes … from $bootstrap-path”, we import the bootstrap
file explicitly at the top of any file that needs classes from it, and
use sass’s extend function instead. this way postcss-scss ignores it
instead of getting confused.