use fixed position for sticky divs in postdetails not absolute
the sticky divs in the postdetails component are especially complex in that there are two of them.
to get the activity div acting appropriately requires taking into account the positioning of the sticky header div above it

in addition, using position: fixed means losing the ability to set the width using just css because it doesnt inherit width from
its container. this means getting the width of a component whose width should match the fixed one, and setting the width in the
component using inline styles, according to that.