Support for prerequisite groups
Add new settings section to Group Settings Tab to add any number of current parent groups as prerequisite groups meaning a person has to join all the prerequisites before they are allowed to join this group.
When looking at the group details page/panel for a group you are not a member of that has prerequisites you see the list of other groups you have to join. If they are open you can join automatically, if they are restricted you have to request to join. If that group has other prereqs then you have to click through to it.