Made it so you can type in topics with hashtags. Small bug(?) introduced... when entering a hashtag, if the only thing entered is the octothorpe, the status message still says "Start by adding a topic" but it is now a selectable item in the list. Pressing return cancels the input.
What's happening is that AsyncCreatableSelect thinks there's an input (which there is, even though we take out the octothorpe) and so it's calling formatOptionLabel on line 141.

The behavior should, at some point, shift to the default behavior of noOptionsMessage on line 138. To do this, I think we have to capture the input before it goes to AsyncCreatableSelect, not quite sure how to do this.

As it stands, this behavior is much better than current behavior and the "bug" is not all that bad.