Show public posts to non-logged in users
When not logged in and trying to view any URL that displays a post, check if the post is public and if so redirect to /post/:id which now displays the PostDetail component. If not public then redirects to /login
For now /public/post/:id also redirects to this URL since we dont want to display the public stream just yet.